+++ .formatter.exs
@@ -3,5 +3,9 @@
   inputs: [
     "{mix,.formatter}.exs",
     "{config,lib,test}/**/*.{ex,exs}"
+  ],
+  locals_without_parens: [
+    edge: 1,
+    edge: 2
   ]
 ]